+++ backend/logic/payments/stripe.js
@@ -1,0 +1,75 @@
+const Stripe = require('stripe')
+const { getLogger } = require('../../utils/logger')
+const { ExternalPayment } = require('../../models')
+const { Sentry } = require('../../sentry')
+
+const log = getLogger('utils.stripe')
+
+const { IS_TEST } = require('../../utils/const')
+const { getConfig } = require('../../utils/encryptedConfig')
+
+/**
+ * Refunds a Stripe payment.
+ *
+ * @param {models.Shop} shop: Shop DB object.
+ * @param {models.Order} order: Order DB object.
+ * @returns {Promise<null|string>} Returns null or the reason for the Stripe failure.
+ * @throws {Error}
+ */
+async function processStripeRefund({ shop, order }) {
+  if (IS_TEST) {
+    log.info('Test environment. Skipping Stripe refund logic.')
+    return null
+  }
+
+  // Load the external payment data to get the payment intent.
+  const externalPayment = await ExternalPayment.findOne({
+    where: {
+      paymentCode: order.paymentCode
+    },
+    attributes: ['payment_code', 'payment_intent']
+  })
+  if (!externalPayment) {
+    throw new Error(
+      `Failed loading external payment with code ${order.paymentCode}`
+    )
+  }
+
+  const paymentIntent = externalPayment.get({ plain: true }).payment_intent
+  if (!paymentIntent) {
+    throw new Error(
+      `Missing payment_intent in external payment with id ${externalPayment.id}`
+    )
+  }
+  log.info('Payment Intent', paymentIntent)
+
+  const shopConfig = getConfig(shop.config)
+
+  // Call Stripe to perform the refund.
+  const stripe = Stripe(shopConfig.stripeBackend)
+  const piRefund = await stripe.refunds.create({
+    payment_intent: paymentIntent
+  })
+
+  const refundError = piRefund.reason
+  if (refundError) {
+    // If stripe returned an error, log it but do not throw an exception.
+    // TODO: fine-grained error handling. Some reasons might be retryable.
+    Sentry.captureException(
+      new Error(
+        `[Shop ${shop.id}] Stripe refund for payment intent ${paymentIntent} failed: ${refundError}`
+      )
+    )
+    log.error(
+      `[Shop ${shop.id}] Stripe refund for payment intent ${paymentIntent} failed: ${refundError}`
+    )
+    return refundError
+  }
+
+  log.info('Payment refunded')
+  return null
+}
+
+module.exports = {
+  processStripeRefund
+}
