+++ backend/utils/shop.js
@@ -77,6 +77,19 @@ function findShopByHostname(req, res, next) {
 }
 
 /**
+ * Middleware to find a shop using `shopId` param
+ * @param {*} req
+ * @param {*} res
+ * @param {*} next
+ */
+const findShop = async (req, res, next) => {
+  const { shopId } = req.params
+  const shop = await Shop.findOne({ where: { id: shopId } })
+  req.shop = shop
+  next()
+}
+
+/**
  * Get a shop's public URL. To be used before a shop is created.
  * @param {string} hostname: Shop's hostname.
  * @param {string} domain: Network domain or localhost on dev environment.
@@ -132,6 +145,7 @@ function getShopDataUrl(shop, networkConfig) {
 
 module.exports = {
   createShop,
+  findShop,
   findShopByHostname,
   getShopDataUrl,
   getShopPublicUrl,
