+++ backend/utils/stripe.js
@@ -2,12 +2,8 @@ const { normalizeDescriptor } = require('@origin/utils/stripe')
 const Stripe = require('stripe')
 const get = require('lodash/get')
 const { getLogger } = require('./logger')
-const { ExternalPayment } = require('../models')
 
 const log = getLogger('utils.stripe')
-
-const { IS_TEST } = require('../utils/const')
-const { getConfig } = require('./encryptedConfig')
 
 /**
  * Forms and returns the webhook object to be used to create
@@ -233,68 +229,10 @@ async function webhookValidation(shop, config, backendUrl) {
   return valid
 }
 
-/**
- * Refunds a Stripe payment.
- *
- * @param {models.Shop} shop: Shop DB object.
- * @param {models.Order} order: Order DB object.
- * @returns {Promise<null|string>} Returns null or the reason for the Stripe failure.
- * @throws {Error}
- */
-async function processStripeRefund({ shop, order }) {
-  if (IS_TEST) {
-    log.info('Test environment. Skipping Stripe refund logic.')
-    return null
-  }
-
-  // Load the external payment data to get the payment intent.
-  const externalPayment = await ExternalPayment.findOne({
-    where: {
-      paymentCode: order.paymentCode
-    },
-    attributes: ['payment_code', 'payment_intent']
-  })
-  if (!externalPayment) {
-    throw new Error(
-      `Failed loading external payment with code ${order.paymentCode}`
-    )
-  }
-
-  const paymentIntent = externalPayment.get({ plain: true }).payment_intent
-  if (!paymentIntent) {
-    throw new Error(
-      `Missing payment_intent in external payment with id ${externalPayment.id}`
-    )
-  }
-  log.info('Payment Intent', paymentIntent)
-
-  const shopConfig = getConfig(shop.config)
-
-  // Call Stripe to perform the refund.
-  const stripe = Stripe(shopConfig.stripeBackend)
-  const piRefund = await stripe.refunds.create({
-    payment_intent: paymentIntent
-  })
-
-  const refundError = piRefund.reason
-  if (refundError) {
-    // If stripe returned an error, log it but do not throw an exception.
-    // TODO: fine-grained error handling. Some reasons might be retryable.
-    log.error(
-      `[Shop ${shop.id}] Stripe refund for payment intent ${paymentIntent} failed: ${refundError}`
-    )
-    return refundError
-  }
-
-  log.info('Payment refunded')
-  return null
-}
-
 module.exports = {
   getWebhookData,
   normalizeDescriptor,
   deregisterWebhooks,
   registerWebhooks,
-  webhookValidation,
-  processStripeRefund
+  webhookValidation
 }
